requirements.txt
psycopg2-binary
requests
python-dotenv
pandas # ยังไม่ใช้